Rename ColumnDefinition.column_index to column_id
Since the index will be the reference and the data, across migrations
where columns might be added/removed over time, rename it to `column_id`
to more accurately refer to that instead.

Having it named `column_index` makes it sound to refer to where in the
`RecordData` vec the value is, which is not the case.